Fix edit bug!
The problem had to do with the SQL query.
It turns out that both tables, Transactions and
Categories, had columns named "_id". When the
SELECT_ALL was called for Transactions, the
Category _id overwrote the Transaction ID,
which is why the row wouldn't update.

I tried a number of approaches to fix this,
but settled on an even more complicated query
that only grabs the `title` column from Categories
instead of `categories.*`.

Also, I converted the Transactions ListView to use
TransactionsArrayAdapter. I built up some code
debt with database code mixed in with display
code, and I am trying to move away from it
by keeping everything in DbHelper or DbContract.